+++ css/passwordManager.css
@@ -1,4 +1,4 @@
-#bitwarden-setup-dialog {
+#manager-setup-dialog {
   width: 100%;
   max-width: 500px;
   border-radius: 8px;
@@ -8,7 +8,7 @@
   background-color: #fff;
 }
 
-.dark-mode #bitwarden-setup-dialog {
+.dark-mode #manager-setup-dialog {
   background-color: rgb(33, 37, 43);
 }
 
@@ -26,10 +26,10 @@
   text-align: center;
 }
 
-#bitwarden-setup-dialog .buttons {
+#manager-setup-dialog .buttons {
   float: right;
 }
 
-#bitwarden-setup-dialog input[type='button'] {
+#manager-setup-dialog input[type='button'] {
   padding: 0.2em;
 }
