+++ index.html
@@ -146,8 +146,8 @@
 		<button id="download-close-button" class="fa fa-times"></button>
 	</div>
 
-	<div class="centered" id="bitwarden-setup-dialog" hidden>
-		<h3 id="bitwarden-setup-heading"></h3>
+	<div class="centered" id="manager-setup-dialog" hidden>
+		<h3 id="manager-setup-heading"></h3>
 		<div class="light-fade" data-string="passwordManagerSetupStep1" data-allowHTML></div>
 		<div class="spacer"></div>
 		<div class="light-fade" data-string="passwordManagerSetupStep2" data-allowHTML></div>
@@ -160,8 +160,8 @@
 		<div class="spacer"></div>
 
 		<div class="buttons">
-			<input type="button" id="bitwarden-setup-disable" data-value="disableAutofill" />
-			<input type="button" id="bitwarden-setup-cancel" data-value="autofillNotNow" />
+			<input type="button" id="manager-setup-disable" data-value="disableAutofill" />
+			<input type="button" id="manager-setup-cancel" data-value="autofillNotNow" />
 		</div>
 
 	</div>
